requirements.txt
fake-useragent
requests
python-dotenv
pyttsx3
TikTokLive
